requirements.txt
annotated-types==0.7.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
coverage==7.6.10
distlib==0.3.9
filelock==3.17.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
identify==2.6.6
idna==3.10
iniconfig==2.0.0
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-cov==6.0.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
SQLAlchemy==2.0.38
tabulate>=0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.1
